# swaggerui/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "OpenAPI standard for Portfolio Tracker",
    "title": "dbservice-api",
    "version": "1.0.4",
    "contact": {
      "name": "API Support",
      "url": "http://www.threefoldsystems.com/support",
      "email": "info@threefooldsystems.com"
    }
  },
  "paths": {
    "/dbservice/api/v1/sys/info/isalive": {
      "get": {
        "summary": "Openshift readiness and liveliness probes",
        "responses": {
          "200": {
            "description": "Readiness and liveliness probe check",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/setup": {
      "post": {
        "summary": "Api call that saves affiliate info and initialises the document store",
        "operationId": "dbsetup",
        "requestBody": {
          "description": "Affiliate schema data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Affiliate"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Setup the document store database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Response"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/migrate": {
      "post": {
        "summary": "Api endpoint that calls the Tradesmiths API and does a migrate to the document store",
        "operationId": "dbmigrate",
        "requestBody": {
          "description": "Affiliate schema data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Affiliate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Migrate to the document store database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Response"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/index": {
      "post": {
        "summary": "Api call that indexes the document store for efficient queries",
        "operationId": "dbindex",
        "responses": {
          "200": {
            "description": "Migrate to the document store database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Affiliate"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/specific": {
      "post": {
        "summary": "Api call that calls the Tradsmiths API and does specific mapping to stock items",
        "operationId": "dbspecific",
        "requestBody": {
          "description": "Stock schema data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Affiliate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Migrate to the document store database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Response"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/affiliates": {
      "get": {
        "summary": "Api call that gets all affiliates",
        "operationId": "dbgetaffiliates",
        "responses": {
          "200": {
            "description": "Get all affiliates",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Response"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/publications/{affiliateid}": {
      "get": {
        "summary": "Api call that gets all affiliates",
        "operationId": "dbgetpublications",
        "parameters": [
          {
            "name": "affiliateid",
            "in": "path",
            "description": "affiliate id (for associated publications)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all publications for an affiliate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Response"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/stocks/{publicationid}": {
      "get": {
        "summary": "Api call that gets all stocks for a publication",
        "operationId": "dbgetstocksforpublication",
        "parameters": [
          {
            "name": "publicationid",
            "in": "path",
            "description": "publication id (for associated stocks)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all stocks for a specific publication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Response"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/stocks/affiliate/{affiliateid}": {
      "get": {
        "summary": "Api call that gets all stocks for an affiliate (all publications)",
        "operationId": "dbgetstocksforaffiliate",
        "parameters": [
          {
            "name": "affiliateid",
            "in": "path",
            "description": "affiliate id (for associated stocks)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all stocks for a specific publication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Response"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/stocks/{bsonid}": {
      "put": {
        "summary": "Api call that updates a specific stock",
        "operationId": "dbupdatestock",
        "parameters": [
          {
            "name": "bsonid",
            "in": "path",
            "description": "bson id (uid object)",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update a stock",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Response"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/watchlist/{bsonid}": {
      "put": {
        "summary": "Api call that updates the watchlist",
        "operationId": "dbupdatewatchlist",
        "parameters": [
          {
            "name": "bsonid",
            "in": "path",
            "description": "bson id (uid object)",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update watchlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Response"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/watchlist/{customerid}": {
      "get": {
        "summary": "Api call that gets a customer's watchlist",
        "operationId": "dbgetwatchlist",
        "parameters": [
          {
            "name": "customerid",
            "in": "path",
            "description": "customer id",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get a watchlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Response"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dbservice/api/v1/prices": {
      "post": {
        "summary": "Api call that updates the stocks with current prices",
        "operationId": "dbupdatestockcurrentprice",
        "responses": {
          "200": {
            "description": "Update stocks with current price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "500": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Response"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": " https://nginx-service-portfoliotracker.e4ff.pro-eu-west-1.openshiftapps.com",
      "variables": {
        "port": {
          "default": "443"
        },
        "basePath": {
          "default": "/"
        }
      },
      "description": "base path and url for microservices"
    }
  ],
  "components": {
    "schemas": {
      "SchemaInterface": {
        "type": "object",
        "properties": {
          "id": {
            "type": "object",
            "format": "int64"
          },
          "lastupdate": {
            "type": "number",
            "format": "int64"
          },
          "metainfo": {
            "type": "string"
          },
          "affiliates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Affiliate"
            }
          },
          "publications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Publication"
            }
          },
          "stocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Stock"
            }
          },
          "watch": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Watchlist"
            }
          }
        }
      },
      "Affiliate": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "object",
            "format": "int64"
          },
          "id": {
            "type": "number",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "Publication": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "object",
            "format": "int64"
          },
          "id": {
            "type": "number",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "affiliateid": {
            "type": "number",
            "format": "int64"
          }
        }
      },
      "Watchlist": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "object",
            "format": "int64"
          },
          "customerid": {
            "type": "number",
            "format": "int64"
          },
          "stocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Stock"
            }
          }
        }
      },
      "Stock": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "object",
            "format": "int64"
          },
          "publicationid": {
            "type": "number",
            "format": "int64"
          },
          "affiliateid": {
            "type": "number",
            "format": "int64"
          },
          "refid": {
            "type": "number",
            "format": "int64"
          },
          "symbol": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "buy": {
            "type": "number",
            "format": "float64"
          },
          "stop": {
            "type": "number",
            "format": "float64"
          },
          "last": {
            "type": "number",
            "format": "float64"
          },
          "change": {
            "type": "number",
            "format": "float64"
          },
          "recommendation": {
            "type": "string"
          },
          "status": {
            "type": "number",
            "format": "int64"
          }
        }
      },
      "Response": {
        "type": "object",
        "properties": {
          "statuscode": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/SchemaInterface"
          }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Entity not found."
      },
      "IllegalInput": {
        "description": "Illegal input for operation."
      },
      "Forbidden": {
        "description": "Not allowed to access this endpoint"
      },
      "GeneralError": {
        "description": "General Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}